# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_readygeneration_readygeneration-backend_pkg_response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.APIError'
      meta:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Meta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  internal_handler.captureLeadRequest:
    properties:
      city:
        type: string
      email:
        type: string
      first_name:
        type: string
      interested_grants:
        items:
          type: string
        type: array
      last_name:
        type: string
      org_name:
        type: string
      org_type:
        type: string
      phone:
        type: string
      quiz_responses:
        additionalProperties: true
        type: object
      source:
        type: string
      state:
        type: string
      utm_campaign:
        type: string
      utm_medium:
        type: string
      utm_source:
        type: string
      zip:
        type: string
    required:
    - email
    type: object
...
  internal_handler.convertLeadRequest:
    properties:
      org_id:
        type: string
    required:
    - org_id
    type: object
...
  internal_handler.generateNarrativeRequest:
    properties:
      application_id:
        type: string
      custom_notes:
        type: string
      section:
        type: string
      word_target:
        type: integer
    required:
    - section
    type: object
...
  internal_handler.updateLeadRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      notes:
        type: string
      org_name:
        type: string
      phone:
        type: string
      score:
        type: integer
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Grant intelligence platform API
  title: ReadyGeneration API
  version: "1.0"
paths:
...
  /admin/leads:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List all leads (admin)
      tags:
      - leads
  /admin/leads/{id}:
    get:
      parameters:
      - description: Lead UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get a lead by ID
      tags:
      - leads
    patch:
      consumes:
      - application/json
      parameters:
      - description: Lead UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update a lead record
      tags:
      - leads
  /admin/leads/{id}/activities:
    get:
      parameters:
      - description: Lead UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get activity timeline for a lead
      tags:
      - leads
  /admin/leads/{id}/convert:
    post:
      consumes:
      - application/json
      parameters:
      - description: Lead UUID
        in: path
        name: id
        required: true
        type: string
      - description: Target organization
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.convertLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Convert a lead to a paying organization
      tags:
      - leads
...
  /applications/{application_id}/narratives:
    get:
      parameters:
      - description: Application UUID
        in: path
        name: application_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List AI narratives for an application
      tags:
      - narratives
...
  /applications/{id}/narratives:
    get:
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get all AI narratives for an application
      tags:
      - applications
...
  /leads:
    post:
      consumes:
      - application/json
      parameters:
      - description: Lead data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.captureLeadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: Capture a new lead (from quiz or landing page)
      tags:
      - leads
...
  /orgs/{org_id}/grants/{grant_id}/narratives:
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization UUID
        in: path
        name: org_id
        required: true
        type: string
      - description: Grant UUID
        in: path
        name: grant_id
        required: true
        type: string
      - description: Generation parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.generateNarrativeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Generate an AI narrative section for a grant application
      tags:
      - narratives
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"